% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chs_result_function.R
\name{chs_c_rate}
\alias{chs_c_rate}
\title{Calculating crude rate.}
\usage{
chs_c_rate(data, by_var = "sex", de_var = "sm_a0100",
  cluster = "JIJUM_CD", strata = c("BOGUN_CD", "dong_type",
  "house_type"), index_subset = NULL, josa_year,
  index_year = 2011:2018, var_cat = 1, weight = "wt", digits = 1)
}
\arguments{
\item{data}{data set.}

\item{by_var}{strata variable.}

\item{de_var}{target(dependent) variable.}

\item{cluster}{cluster.}

\item{strata}{strate.}

\item{index_subset}{denominator condition by health indicators.}

\item{josa_year}{the year of data.}

\item{index_year}{the years in which the target variable was investigated.}

\item{weight}{weight.}

\item{digits}{integer indicating the number of decimal places.#'}
}
\value{
crude rate of health indicators
}
\description{
chs_c_rate function will calculate crude rate of health indicators.
}
