% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chs_result_function.R
\name{chs_c_rate_by}
\alias{chs_c_rate_by}
\title{Calculating crude rate by the stratification variable.}
\usage{
chs_c_rate_by(data, by_var, ...)
}
\arguments{
\item{data}{data set.}

\item{by_var}{stratification variable.}
}
\value{
crude rate by stratificaton variable.
}
\description{
chs_c_rate_by function will calculate crude rate of health indicators by the stratification variable.
}
