% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_district_variable.R
\encoding{UTF8}
\name{code_district_daegu}
\alias{code_district_daegu}
\title{coding district varible}
\usage{
code_district_daegu(data_set, var_dong = "dong_p", var_district = "district")
}
\arguments{
\item{data_set}{community health survey raw data}

\item{var_dong}{dong variable of chs data}

\item{var_district}{district variable (new)}
}
\value{

}
\description{
code_district_daegu functions will code district variable.
}
