% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_district_variable.R
\encoding{UTF8}
\name{code_district_ulsan}
\alias{code_district_ulsan}
\title{coding district varible}
\usage{
code_district_ulsan(data_set, var_dong = "dong_p",
  var_district = "district")
}
\arguments{
\item{data_set}{community health survey raw data}

\item{var_dong}{dong variable of chs data}

\item{var_district}{district variable (new)}
}
\description{
code_district_ulsan functions will code district variable.
}
