% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_coding.R
\encoding{UTF8}
\name{data_coding}
\alias{data_coding}
\title{coding chs index.}
\usage{
data_coding(data_set, year, smoking = TRUE)
}
\arguments{
\item{data_set}{raw data of community health survey.}

\item{year}{the year investigated.}

\item{smoking}{whether smoking section variables are coded}
}
\description{
data_coding function will code health indicators using raw data of community health survey.
}
