% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_coding2.R
\encoding{UTF8}
\name{data_coding2}
\alias{data_coding2}
\title{coding chs index.}
\usage{
data_coding2(data, year)
}
\arguments{
\item{year}{the year investigated.}

\item{data_set}{raw data of community health survey.}
}
\value{

}
\description{
data_coding2 function will code health indicators using raw data of community health survey.
}
