% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_coding_meduse.R
\encoding{UTF8}
\name{data_coding_meduse}
\alias{data_coding_meduse}
\title{coding chs index.}
\usage{
data_coding_meduse(data_set, year)
}
\arguments{
\item{data_set}{raw data of community health survey.}

\item{year}{the year investigated.}
}
\description{
data_coding_meduse function will code health indicators using raw data of community health survey.
}
