% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_read.R
\name{dataread_chs_raw}
\alias{dataread_chs_raw}
\title{Reading chs data.}
\usage{
dataread_chs_raw(year, filename, data3 = FALSE)
}
\arguments{
\item{year}{the year investigated.}

\item{filename}{Path to the xls/xlsx file.}

\item{data3}{TRUE to use third data sheet, FALSE to use only two data sheet.}
}
\description{
dataread_chs_raw function will read raw data of community health survey.
}
